fix(render): widen InstDrawPerPlayer pointer fields to fix 64-bit truncation
ptrCommandList/ptrColorLayout/ptrDeltaArray were stored as u32/int and
written via truncating (u32)/(int) casts from already-wide pointer sources
(ModelHeader.ptrCommandList is uintptr_t, ModelHeader.ptrColors and
ModelAnim.ptrDeltaArray are u32*). On 64-bit hosts this dropped the upper
32 bits, producing garbage pointers at every render-time read and crashing
RenderBucket_CopyScratchColorCache. Widened all three fields (and the
RenderBucket_GetFrame out-param/local that feeds ptrDeltaArray) to
uintptr_t, matching the established ModelHeader.ptrCommandList pattern;
all affected offsets are CTR_STATIC_ASSERT_LAYOUT-gated so 32-bit builds
are unaffected.

Also fixed ModelHeader.unk3 (aliases ModelAnim.ptrDeltaArray for static
models per source comment) to go through Reloc64_Resolve like its sibling
fields instead of being copied as a raw disc value.

Verified: macOS arm64 build clean (473 vs 487 warnings pre-fix), and the
targeted crash no longer reproduces -- the binary now runs deterministically
further into the render path before hitting a separate, unrelated bug in
the OT/primitive-link bridge (RenderBucket_LinkPrimRaw), tracked as a
follow-up.

+++ game/232/AH_WarpPad.c
@@ -84,8 +84,8 @@ void AH_WarpPad_SetNumModelData(struct Instance *inst, struct ModelHeader *mh)
 {
 	struct InstDrawPerPlayer *idpp = INST_GETIDPP(inst);
 
-	idpp[0].ptrCommandList = (u32)mh->ptrCommandList;
-	idpp[0].ptrColorLayout = (u32)mh->ptrColors;
+	idpp[0].ptrCommandList = mh->ptrCommandList;
+	idpp[0].ptrColorLayout = (uintptr_t)mh->ptrColors;
 	idpp[0].ptrTexLayout = mh->ptrTexLayout;
 	idpp[0].ptrCurrFrame = mh->ptrFrameData;
 }
